backend/requirements.txt
langchain
langchain-community
pypdf
beautifulsoup4
requests
lxml
python-multipart
sentence-transformers
torch
fastapi
uvicorn
google-generativeai
langchain-text-splitters
python-dotenv
requests
chromadb
whoosh
pydantic
